requirements.txt
flask==3.1.2
pyserial==3.5
pynput==1.7.6
pywebview==5.3
click==8.3.1
Jinja2==3.1.6
Werkzeug==3.0.1
itsdangerous==2.2.0
